refactor bindings
