+++ rabix-bindings/src/main/java/org/rabix/bindings/protocol/draft2/Draft2Bindings.java
@@ -1,21 +1,30 @@
-package org.rabix.bindings;
+package org.rabix.bindings.protocol.draft2;
 
 import java.io.File;
 import java.util.List;
 import java.util.Set;
 
+import org.rabix.bindings.BindingException;
+import org.rabix.bindings.Bindings;
+import org.rabix.bindings.CommandLineBuilder;
+import org.rabix.bindings.DocumentReferenceResolver;
+import org.rabix.bindings.ProtocolJobHelper;
+import org.rabix.bindings.ProtocolProcessor;
+import org.rabix.bindings.ProtocolTranslator;
+import org.rabix.bindings.ProtocolType;
+import org.rabix.bindings.ProtocolValueOperator;
+import org.rabix.bindings.RequirementProvider;
+import org.rabix.bindings.ResultCollector;
 import org.rabix.bindings.filemapper.FileMapper;
 import org.rabix.bindings.model.FileValue;
 import org.rabix.bindings.model.Job;
 import org.rabix.bindings.model.dag.DAGNode;
 import org.rabix.bindings.model.requirement.Requirement;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.rabix.bindings.protocol.draft2.helper.Draft2ProtocolJobHelper;
+import org.rabix.bindings.protocol.draft2.resolver.Draft2DocumentReferenceResolver;
 
-public class BindingsImpl implements Bindings {
+public class Draft2Bindings implements Bindings {
 
-  private static final Logger logger = LoggerFactory.getLogger(BindingsImpl.class);
-  
   private final ProtocolType protocolType;
   
   private final ResultCollector resultCollector;
@@ -27,21 +36,16 @@ public class BindingsImpl implements Bindings {
   private final ProtocolJobHelper protocolJobHelper;
   private final DocumentReferenceResolver documentReferenceResolver;
   
-  public BindingsImpl(ProtocolType protocolType) throws BindingException {
-    try {
-      this.protocolType = protocolType;
-      this.resultCollector = ResultCollectorFactory.create(protocolType);
-      this.protocolProcessor = ProtocolProcessorFactory.create(protocolType);
-      this.commandLineBuilder = CommandLineBuilderFactory.create(protocolType);
-      this.valueOperator = ProtocolValueOperatorFactory.create(protocolType);
-      this.protocolTranslator = ProtocolTranslatorFactory.create(protocolType);
-      this.requirementProvider = RequirementProviderFactory.create(protocolType);
-      this.protocolJobHelper = ProtocolJobHelperFactory.create(protocolType);
-      this.documentReferenceResolver = DocumentReferenceResolverFactory.create(protocolType);
-    } catch (BindingException e) {
-      logger.error("Failed to create Bindings for type " + protocolType, e);
-      throw new BindingException("Failed to create Bindings for type " + protocolType);
-    }
+  public Draft2Bindings() throws BindingException {
+    this.protocolType = ProtocolType.DRAFT2;
+    this.resultCollector = new Draft2ResultCollector();
+    this.protocolProcessor = new Draft2ProtocolProcessor();
+    this.commandLineBuilder = new Draft2CommandLineBuilder();
+    this.valueOperator = new Draft2ProtocolValueExtractor();
+    this.protocolTranslator = new Draft2ProtocolTranslator();
+    this.requirementProvider = new Draft2RequirementProvider();
+    this.protocolJobHelper = new Draft2ProtocolJobHelper();
+    this.documentReferenceResolver = new Draft2DocumentReferenceResolver();
   }
   
   @Override
